int to llint

+++ pi.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <math.h>
 
-double calcPI(int n) {
+double calcPI(long long int n) {
 	if (n == 1) return 4;
 	return 4*pow(-1,n+1)*(1/(double)(2*n-1))+calcPI(n-1);
 }
@@ -17,7 +17,7 @@ int main(int argc, char *argv[]) {
     if (strcmp(argv[1], "--help") == 0) {goto exit;}
 
     double pi = 0;
-    int n = atoi(argv[1]);
+    long long int n = atoll(argv[1]);
     
 	printf("calcuating pi...\n");
 	pi = calcPI(n);
